5-1 and 5-2 complete
There's something funky about 5-2 where it fails or delivers odd inaccuracies for certain large or high precision floats. Need to investigate more of how float storage works in order to fix.

+++ KR5/KR5-1to2/getter_tester.h
@@ -10,4 +10,4 @@ char buf[BUFSIZE]; /* buffer for ungetch */
 int bufp = 0; /* next free position in buf */
 
 int getint(int *pn);
-float getfloat(float *pf);
+int getfloat(float *pf);
